+++ src/app/Peerly/types.ts
@@ -4,3 +4,5 @@ export interface APIError {
   status_code: number;
   data: null;
 }
+
+export type BadgeType = 'platinum' | 'gold' | 'silver' | 'bronze' | 'basicUser';
